Refactor duke-classfile types into cohesive modules
