+++ crates/duke-classfile/src/constant_pool.rs
@@ -1,0 +1,101 @@
+/// Newtype wrapper for constant pool indices (1-based per JVM spec).
+///
+/// # Examples
+///
+/// ```
+/// use duke_classfile::types::CpIndex;
+///
+/// let index = CpIndex(42);
+/// assert_eq!(index.0, 42);
+/// ```
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
+pub struct CpIndex(pub u16);
+
+/// All constant pool entry kinds defined in JVM SE 21 (§4.4).
+#[derive(Debug, Clone, PartialEq)]
+pub enum CpEntry {
+    /// Tag 1
+    Utf8(String),
+    /// Tag 3
+    Integer(i32),
+    /// Tag 4
+    Float(f32),
+    /// Tag 5 — occupies two slots; next slot will be `None`
+    Long(i64),
+    /// Tag 6 — occupies two slots; next slot will be `None`
+    Double(f64),
+    /// Tag 7
+    Class {
+        /// Index to a `CONSTANT_Utf8` structure representing a valid binary class or interface name.
+        name_index: CpIndex,
+    },
+    /// Tag 8
+    String {
+        /// Index to a `CONSTANT_Utf8` structure representing the string's value.
+        string_index: CpIndex,
+    },
+    /// Tag 9
+    Fieldref {
+        /// Index to a `CONSTANT_Class` structure.
+        class_index: CpIndex,
+        /// Index to a `CONSTANT_NameAndType` structure.
+        name_and_type_index: CpIndex,
+    },
+    /// Tag 10
+    Methodref {
+        /// Index to a `CONSTANT_Class` structure.
+        class_index: CpIndex,
+        /// Index to a `CONSTANT_NameAndType` structure.
+        name_and_type_index: CpIndex,
+    },
+    /// Tag 11
+    InterfaceMethodref {
+        /// Index to a `CONSTANT_Class` structure.
+        class_index: CpIndex,
+        /// Index to a `CONSTANT_NameAndType` structure.
+        name_and_type_index: CpIndex,
+    },
+    /// Tag 12
+    NameAndType {
+        /// Index to a `CONSTANT_Utf8` structure representing a valid unqualified name.
+        name_index: CpIndex,
+        /// Index to a `CONSTANT_Utf8` structure representing a valid field or method descriptor.
+        descriptor_index: CpIndex,
+    },
+    /// Tag 15 — method handles (JSR 292)
+    MethodHandle {
+        /// The kind of method handle.
+        reference_kind: u8,
+        /// The reference index for the method handle.
+        reference_index: CpIndex,
+    },
+    /// Tag 16
+    MethodType {
+        /// Index to a `CONSTANT_Utf8` structure representing a method descriptor.
+        descriptor_index: CpIndex,
+    },
+    /// Tag 17
+    Dynamic {
+        /// An index into the bootstrap method table.
+        bootstrap_method_attr_index: u16,
+        /// Index to a `CONSTANT_NameAndType` structure.
+        name_and_type_index: CpIndex,
+    },
+    /// Tag 18
+    InvokeDynamic {
+        /// An index into the bootstrap method table.
+        bootstrap_method_attr_index: u16,
+        /// Index to a `CONSTANT_NameAndType` structure.
+        name_and_type_index: CpIndex,
+    },
+    /// Tag 19 — Java 9+ module system
+    Module {
+        /// Index to a `CONSTANT_Utf8` structure representing a module name.
+        name_index: CpIndex,
+    },
+    /// Tag 20 — Java 9+ module system
+    Package {
+        /// Index to a `CONSTANT_Utf8` structure representing a package name.
+        name_index: CpIndex,
+    },
+}
